% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreedToColor.R
\name{BreedToColor}
\alias{BreedToColor}
\title{Function used in app. Returns colorblends based on CB input as four letter 
abbrivations}
\usage{
BreedToColor(xx, rgbw = "DLWH", lim_white = 0.9, blackBreed = "P")
}
\arguments{
\item{xx}{vector of length n with crossbreeds defined by four letter abbrivations.}

\item{rgbw}{four letter string defining which purebreeds that should be 
represented by colors red, green, blue and white respectively.}

\item{lim_white}{scalar. Defines the amount of "grey" in the white color}

\item{blackBreed}{One letter abbreviation for which all breedcombinations
containing this breed will get black color.}
}
\value{
A vector of length n containing color definitions as returned by rgb()
}
\description{
Function used in app. Returns colorblends based on CB input as four letter 
abbrivations
}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
