% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreedToMat.R
\name{BreedToMat}
\alias{BreedToMat}
\title{Function for redefining breed combinations as 4 letter abbreviations (example 'DDLW') 
to numerical vectors (\eqn{\Delta}'s)}
\usage{
BreedToMat(Bvec, Breeds = "DHLPW")
}
\arguments{
\item{Bvec:}{a vector of length n, each string represents a letter abbrivation for 
on individual.}

\item{Breeds:}{The (m number of) breeds to be included in \eqn{\Delta}'s, either as one string element
("DHLPW" which is the default) or as a vector defining the breeds ("c('D','H','L','P','W')").}
}
\value{
A matrix of size n x m.
}
\description{
Function for redefining breed combinations as 4 letter abbreviations (example 'DDLW') 
to numerical vectors (\eqn{\Delta}'s)
}
\examples{
BreedToMat(c('DDLW','HHLW'))
BreedToMat(c('DLDW','HHLW'))
BreedToMat(c('DLDW','HHLW'),'abDH')
}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
