% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.ind.R
\name{class.ind}
\alias{class.ind}
\title{Function for making an indicator matrix based on vector of factors/ levels}
\usage{
class.ind(x)
}
\arguments{
\item{x}{Vector of length n, which might be transformed to factor by as.factor(x)}
}
\value{
A matrix of size n x m, where m is the number of levels in x (as.factor(x)). 
The matrix has column names as given by levels(as.factor(x)), and contains values 0 or 1
}
\description{
Function for making an indicator matrix based on vector of factors/ levels
}
\examples{
class.ind(as.factor(rep(1:4,each = 5)))
}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
