% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dist_eval_func.R
\name{Dist_eval_func}
\alias{Dist_eval_func}
\title{Function for evaluating Kullback Leibler distances and square losses from a
matrix where rows contain soft predictions and a vector defining the true 
breeds. Used for ecvaluation of simulated data in Vinje et.al.....}
\usage{
Dist_eval_func(Soft, TrueBreed, PBs = c("D", "H", "L", "P", "W"))
}
\arguments{
\item{Soft}{matrix of size n x q where each row represent a soft prediction of 
an individual.}

\item{TrueBreed}{vector of length n containing the true breed combinations as four letter abbrivations.}

\item{PBs:}{Vector of letters defining which PB's that should be included as 
possible PB grandparents. Of the form  "c('D','H','L','P','W')". Only capital letters 
'D','H','L','P','W' are allowed. Default value is c('D','H','L','P','W'). PBs must
include all PB's defined in names of "Soft" and possibly more.}
}
\value{
A data frame of size n x 4 with the four elements:\cr
-KL_dist: Kullback leibler distances between soft and true predications.\cr
-Sq_loss: Square losses between soft and true predications\cr
-KL_hardPred: Hard prediction as the crossbreed combination with smallest KL distance to soft prediction\cr
-Sq_hardPred Hard prediction as the crossbreed combination with smallest squared loss to soft prediction\\cr
}
\description{
Function for evaluating Kullback Leibler distances and square losses from a
matrix where rows contain soft predictions and a vector defining the true 
breeds. Used for ecvaluation of simulated data in Vinje et.al.....
}
\examples{
Dist_eval_func(RF_Sim_predictions$`TrainP+`,rownames(RF_Sim_predictions$`TrainP+`))
}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
