% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributionplot_func.R
\name{Distributionplot_func}
\alias{Distributionplot_func}
\title{A function for making score plot comparing PLS and PCA scores in Vinje et al.}
\usage{
Distributionplot_func(
  image_path = NULL,
  TrueBreed,
  SoftPreds,
  BreedNames = c("Duroc", "Hampshire", "Landrace", "Pietrain", "Large white"),
  MethodNames = c("Random Forest", "PLSR", "PLS-QDA"),
  dens_adj = c(1, 1, 200)
)
}
\arguments{
\item{image_path}{(optional). Path for saving figure as png-format. If NULL 
(default), windows() is called and figure is shown in new windows.}

\item{TrueBreed}{Vector of length n with True Breeds as four letter abbreviations.}

\item{SoftPreds}{data.frame with k (typically 3) elements with soft predictions
of the individuals defined in "TrueBreed"}

\item{BreedNames}{in full not one letter abbrivations}

\item{MethodNames}{names for the k methods (in full) represented in "SoftPreds"}

\item{dens_adj}{vector of length k passed to "density(...,adjust = dens_adj[k])"}
}
\value{
A plot (or png file) with four panels. See Vinje et.al.... for further description.
}
\description{
A function for making score plot comparing PLS and PCA scores in Vinje et al.
}
\examples{
Distributionplot_func(TrueBreed = rownames(RF_Sim_predictions$`TrainP+`),
SoftPreds = data.frame(I(RF_Sim_predictions$`TrainP+`)),
I(PLSR_Sim_predictions$`TrainP+`)),MethodNames = c("Random Forest","PLSR"))
}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
