% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ellipse_func.R
\name{Ellipse_func}
\alias{Ellipse_func}
\title{Function used in app. Returns an ellipse element with isovalue defined by the 
log posteriori value for "Unknown breed"}
\usage{
Ellipse_func(log_unknown, mu, Sigma, log_prior)
}
\arguments{
\item{log_unknown}{A posteriori value (product of prior x likelihood) for unknown 
breed at the log scale.}

\item{mu}{Expected value for the CB in question.}

\item{Sigma}{Covariance matrix for the CB in question}

\item{log_prior}{Prior probability for CB in question at log scale.}
}
\value{
a ellpise3d object
}
\description{
Function used in app. Returns an ellipse element with isovalue defined by the 
log posteriori value for "Unknown breed"
}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
