% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_Models_func.R
\name{Fit_Models_func}
\alias{Fit_Models_func}
\title{A function for fitting PLS, PCA and Random forest models based on PB-data}
\usage{
Fit_Models_func(PB_data)
}
\arguments{
\item{PB_data}{A data frame containing SNP-data for n = 4014 pure breed pigs. PB_data has elements:\cr
- Breed (4 letter abbrivation for real breed) 
- ID. Identification of the different individuals. 
- XX: Matrix of size (4014 x 23070) containing 0-1-2 coding for the 23070 SNp's. 
The column names are the SNP-names. Row names as breed combinations.}
}
\value{
Three objects Mod_pls, Mod_pca and Classififier_RF, all list with two elements:
"TrainP+" and "TrainP-", containing models where Pietrain data is used ("TrainP+") or
omitted ("TrainP-"). \cr
PLS- and PCA- models are fitted by functions plsr() and pcr() from 
the "pls" package.\cr 
The random forest model is fitted via the randomForest() in 
the "randomForest" package. In random forest the weights argument is used in order to
increase the weight of Hampshire observations.\cr
The three model objects saved in the "CrossBreedPrediction" 
package are the same objects as fitted by Fit_Models_func(PB_data), with slight
modifications (i.e. raw data are deleted, and a new element "Breed" is added.)
}
\description{
A function for fitting PLS, PCA and Random forest models based on PB-data
}
\examples{
Data <- Load_012data_func()
attach(Data)
Fit_Models_func(Models$PB_data)
}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
