% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL_dist_CBpred_func.R
\name{KL_dist_CBpred_func}
\alias{KL_dist_CBpred_func}
\title{Function for calculating Kullback Leibler divergence between real and predicted 
CB combinations}
\usage{
KL_dist_CBpred_func(Real, Predicted, pred_min = 10^(-10))
}
\arguments{
\item{Real:}{Vector of length m (number of possible PB), with proportions of each 
PB as quarters, i.e. assuming grandparents to be PB}

\item{Predicted:}{Vector of length m (number of possible PB), with predicted (soft)
proportions for each PB. The order of PB's in Real and Predicted must be the same.}

\item{pred_min:}{Elements <0 or >1 in Predicted are inflated to 0 + pred_min or truncated 
to 1 - pred_min.}
}
\value{
a scalar. The Kullback Leibler divergense assuming that 4 x Real is 
multinomal distributed, with true probability parameter Real, and predicted probability
parameter Predicted.
}
\description{
Function for calculating Kullback Leibler divergence between real and predicted 
CB combinations
}
\examples{
To come..
}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
