% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedTo012_func.R
\name{pedTo012_func}
\alias{pedTo012_func}
\title{Function for reading DNA data (GeneSeek 50K (Illumina) SNP chip) from files (.ped or .txt)}
\usage{
pedTo012_func(ped, na_val = NA)
}
\arguments{
\item{ped:}{A data frame of  size n x 2m, where n is the number of rows (individuals) 
in the ped file, and m is the number of SNP's in the map file (23 070 in the PigBreedPrediction_map)}

\item{na_val:}{Value to put in SNP's with missing data. Default is NA, might be set to 0.}
}
\value{
matrix size n x m, where n is the number of rows (individuals) 
in the ped file, and m is the number of SNP's in the map file 
(23 070 in the PigBreedPrediction_map), use 
head(PigBreedPrediction_map) to view first entries of map file. Column names 
according to SNP-names and Rownames according to ID
}
\description{
Function for reading DNA data (GeneSeek 50K (Illumina) SNP chip) from files (.ped or .txt)
}
\examples{
ped <- read_ped_data('data/Ped_example.txt',ID_column = 'ID',nrows = -1)
ped012 <- pedTo012_func(ped,na_val = 0)
}
\references{
Vinje,H.......
}
\author{
Hilde Brustad, Hilde Vinje & Lars Erik Gangsei
}
