% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PigBreedPrediction_map.R
\docType{data}
\name{PigBreedPrediction_map}
\alias{PigBreedPrediction_map}
\title{Genetic map used in the PigBreedPrediction project.
based on the map for a GeneSeek 50K (Illumina) SNP chip}
\format{
Data frame of size 23070 x 5, with the 5 columns:\cr
- chrm: Chromosome number, integer 1-18\cr
- snp: SNP name, character\cr
- unknown: integer 0 (not in use)\cr
- bp: base position of SNP at chromosome\cr
- dom: dominent base at SNP in the PB animals used in the PigBreedPrediction project\cr
}
\usage{
data(PigBreedPrediction_map)
}
\description{
Genetic map used in the PigBreedPrediction project.
based on the map for a GeneSeek 50K (Illumina) SNP chip
}
\examples{
head(PigBreedPrediction_map)

}
\references{
Vinje et.al. 2023
}
\keyword{datasets,}
\keyword{genetic}
\keyword{map}
