% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosteriorToSoft_func.R
\name{PosteriorToSoft_func}
\alias{PosteriorToSoft_func}
\title{Function to go from a "soft" to hard prediction for crossbreed pigs.}
\usage{
PosteriorToSoft_func(Posteriors, PBs = c("D", "H", "L", "P", "W"))
}
\arguments{
\item{Posteriors:}{Matrix (or numeric vector), with posterior probabilities for different 
crossbreed combinations defined b colomnnames (or names if vector).}

\item{PBs:}{Vector of letters defining which PB's that should be included as 
possible PB grandparents. Of the form  "c('D','H','L','P','W')". Only capital letters 
'D','H','L','P','W' are allowed. Default value is c('D','H','L','P','W'). PBs must
include all PB's defined in names of "Posterior" and possibly more.}
}
\value{
A matrix of dimentions dim(Posterior)[1] x length(PBs) defining soft breed prdictions
based on the posterior given as input. Unknown breed not included.
}
\description{
Function to go from a "soft" to hard prediction for crossbreed pigs.
}
\examples{
To come
}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
