% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pred_func.R
\name{Pred_func}
\alias{Pred_func}
\title{Function used in app. Returns (hard) predictions for the different individuals}
\usage{
Pred_func(PriorDist, LogLike)
}
\arguments{
\item{PriorDist}{named vector with prior probabilities for different
breedcombinations. Length m}

\item{LogLike}{matrix of size n x m with likelihoods for the n individuals to
be classified crossed over the m possible breed combinations. Column names
as four letter abbrivations defing the possible breeds.}
}
\value{
A vector of length n containing four letter abbrivations for hard
predictions of breed combinations.
}
\description{
Function used in app. Returns (hard) predictions for the different individuals
}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
