% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictCB_func.R
\name{PredictCB_func}
\alias{PredictCB_func}
\title{Function for predicting cross breed combination based on DNA data (GeneSeek 50K (Illumina) SNP chip). from files (.ped or .txt)}
\usage{
PredictCB_func(
  DNA = NULL,
  Predictions = NULL,
  Train = "TrainP+",
  alpha0 = 73.58105,
  Unkn_LogLike = NULL,
  PriorDist = NULL,
  pred_min = 10^(-10)
)
}
\arguments{
\item{DNA:}{One of the following: \cr 
- A list with elementsts: 1) Scores: PLS-scores, 2) RF_pred: soft predictions from Random Forest
and 3) PLSR_pred: soft predictions from PLSR. 
- matrix (n x m) with 0-1-2 coding of SNP's to be evaluated\cr
- data.frame (n x 2m) with "A", "C", "G", "T" coding of SNP's to be evaluated\cr
- a path to a txt - file with ped data possible to read with "read_ped_func()"}

\item{Train:}{Either 'TrainP+' (default), or 'TrainP-', indicating if training data including or
omitting PB pietrain is to be used.}

\item{alpha0:}{Deafault value is 73.58105, i.e. the free parameter \eqn{\alpha_0} used 
in calculation of \eqn{V(\theta)}.}

\item{Unkn_LogLike:}{the log likelihood for the uniform distribution for unknown breed in PLS-QDA. 
Defaults value is NULL, for which the value is calculated based on the m dimentional
space spanned bu the scores in the PLS-model for PB's.}

\item{PriorDist:}{"NULL" (default) for which a totally flat prior is applied, 
'informative' for which the informative prior in Vinje et.al.#' is applied, 
or a scalar >0 and <1 defining the prior probability for class "Unknown" 
and a flat prior is applied to 
the rest of the classes,}

\item{pred_min:}{Parameter to be passed to "KL_dist_CBpred_func()". Default to 10^(-10).}
}
\value{
a list bla bla
}
\description{
Function for predicting cross breed combination based on DNA data (GeneSeek 50K (Illumina) SNP chip). from files (.ped or .txt)
}
\examples{
To come
}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
