% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior_func.R
\name{Prior_func}
\alias{Prior_func}
\title{Function used in app. Returns the prior distribution as, where the support 
for Unknown breed and informative prior might be manipiulated by the input}
\usage{
Prior_func(delta_i = 0, delta_UK = 0, Train = "P+")
}
\arguments{
\item{delta_i}{scalar in interval [0 1]. delta_i = 0 is no weight to informative 
prior, whereas delta_i = 1 is the informative prior.}

\item{delta_UK}{scalar in interval [-1 1]. delta_UK = -1 is zero prior 
probability for unknown breed, whereas  delta_UK = 1 is prior probability 1 for
unknown breed.}

\item{Train}{Either 'P+' or 'P-', i.e. which training set to use.}
}
\value{
a named vector (possible breedcombinations) with prior probabilities.
}
\description{
Function used in app. Returns the prior distribution as, where the support 
for Unknown breed and informative prior might be manipiulated by the input
}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
