% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ped_data.R
\name{read_ped_data}
\alias{read_ped_data}
\title{Function for reading DNA data (GeneSeek 50K (Illumina) SNP chip) from files (.ped or .txt)}
\usage{
read_ped_data(ped_path, ID_column = NULL, nrows = -1)
}
\arguments{
\item{ped_path}{Path to "ped" file. The ped file must have column names in line 
with snp names in PigBreedPrediction_map}

\item{ID_column}{(optional) Column names in ped file used as ID.}

\item{nrows}{(optional) pased to "read.table()". integer: the maximum number of rows to 
read in. Negative and other invalid values are ignored. If -1 (default) all rows
are read in.}
}
\value{
A data frame of  size n x 2m, where n is the number of rows (individuals) 
in the ped file, and m is the number of SNP's in the map file 
(23 070 in the PigBreedPrediction_map), use 
head(PigBreedPrediction_map) to view first entries of map file
}
\description{
Function for reading DNA data (GeneSeek 50K (Illumina) SNP chip) from files (.ped or .txt)
}
\examples{
ped <- read_ped_data('data/Ped_example.txt',ID_column = 'ID',nrows = -1)
ped[,1:10]
}
\references{
Vinje,H.......
}
\author{
Hilde Brustad, Hilde Vinje & Lars Erik Gangsei
}
