% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scoreplot_func.R
\name{Scoreplot_func}
\alias{Scoreplot_func}
\title{A function for making score plot comparing PLS and PCA scores in Vinje et al.}
\usage{
Scoreplot_func(image_path = NULL)
}
\arguments{
\item{image_path}{(optional). Path for saving figure as png-format. If NULL 
(default), windos() is called and figure is shown in new windows.}
}
\value{
A plot (or png file) with four panels. See Vinje et.al.... for further description.
}
\description{
A function for making score plot comparing PLS and PCA scores in Vinje et al.
}
\examples{
Scoreplot_func()
}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
