% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoftToHard_func.R
\name{SoftToHard_func}
\alias{SoftToHard_func}
\title{Function to go from a "soft" to hard prediction for crossbreed pigs.}
\usage{
SoftToHard_func(
  Soft,
  method = c("Sq_loss", "KL_dist"),
  PBs = c("D", "H", "L", "W", "P"),
  pred_min = 10^(-10)
)
}
\arguments{
\item{Soft:}{Matrix (or numeric vector), with soft predictions. Each row represents
one individual. Colnames (or names if vector) defines PB's with softbreeds.}

\item{method:}{Eiher 'Sq_loss' or'KL_dist', defining the loss function to be used.}

\item{PBs:}{Vector of letters defining which PB's that should be included as 
possible PB grandparents. Of the form  "c('D','H','L','P','W')". Only capital letters 
'D','H','L','P','W' are allowed. Default value is c('D','H','L','P','W'). PBs must
include all PB's defined in "Soft" and possibly more.}

\item{pred_min:}{Passed to "KL_dist_CBpred_func()". Elements <0 or >1 in 
probability vector to be evaluated for Kullbach-Leibler divergenses are inflated to 
0 + pred_min or truncated to 1 - pred_min.}
}
\value{
A vector of length equal to dim(Soft)[1] with four letter abbreviations, containg 
"hard predictions" for cross breed combinations.
}
\description{
Function to go from a "soft" to hard prediction for crossbreed pigs.
}
\examples{
To come
}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
