% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_func_df.R
\name{sum_func_df}
\alias{sum_func_df}
\title{Function used for function "xtable" in Vinje et.al.....}
\usage{
sum_func_df(df, return_order = c(1, 3, 2, 4))
}
\arguments{
\item{df}{data frame whose elements are either numerical or logical}

\item{return_order}{vector defining order of returned elements}
}
\value{
a vector of length dim(df)[2] containing mean +/- standard deviations
for numerical elements and the sum and %-age in() of "TRUE"'s for logical elements
}
\description{
Function used for function "xtable" in Vinje et.al.....
}
\examples{
Dist_mat <- Dist_eval_func(RF_Sim_predictions$`TrainP+`,rownames(RF_Sim_predictions$`TrainP+`))
sum_func_df(Dist_mat)

}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
