% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_func_hard.R
\name{sum_func_hard}
\alias{sum_func_hard}
\title{Function used for function "xtable" in Vinje et.al.....}
\usage{
sum_func_hard(x)
}
\arguments{
\item{x}{logical vector}

\item{digits}{passed direcetely to "format(...,digits = digits)"}
}
\value{
a string with "percentage of x is TRUE (n = length(x))"
}
\description{
Function used for function "xtable" in Vinje et.al.....
}
\examples{
sum_func_hard(rnorm(10)<0)
sum_func_hard(rnorm(100)<0)
sum_func_hard(rnorm(10^5)<0)

}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
