% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_func_soft.R
\name{sum_func_soft}
\alias{sum_func_soft}
\title{Function used for function "xtable" in Vinje et.al.....}
\usage{
sum_func_soft(x, digits = 2, scientific = 15)
}
\arguments{
\item{x}{numerical vector}

\item{digits}{passed direcetely to "format(...,digits = digits)"}

\item{scientific}{passed direcetely to "format(...,scientific = digits)"}
}
\value{
a string with "mean(x) +/- sd(x)"
}
\description{
Function used for function "xtable" in Vinje et.al.....
}
\examples{
sum_func_soft(rnorm(10))
sum_func_soft(rnorm(100))
sum_func_soft(rnorm(10^5))

}
\references{
Vinje,H.......
}
\author{
Hilde Vinje & Lars Erik Gangsei
}
